% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_blue}
\alias{cei_blue}
\title{CEI blue colour from style guide}
\format{
length 1 character vector
}
\usage{
cei_blue
}
\description{
CEI blue colour from style guide
}
\keyword{datasets}
