% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colour_10}
\alias{cei_colour_10}
\title{CEI colour palette with ten colours}
\format{
list
}
\usage{
cei_colour_10
}
\description{
CEI colour palette with ten colours
}
\keyword{datasets}
