% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ceiglobal.R
\docType{data}
\name{cei_colour_2}
\alias{cei_colour_2}
\title{CEI colour palette with two colours}
\format{
list
}
\usage{
cei_colour_2
}
\description{
CEI colour palette with two colours
}
\keyword{datasets}
