% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colour_3}
\alias{cei_colour_3}
\title{CEI colour palette with three colours}
\format{
list
}
\usage{
cei_colour_3
}
\description{
CEI colour palette with three colours
}
\keyword{datasets}
