% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colour_4}
\alias{cei_colour_4}
\title{CEI colour palette with four colours}
\format{
list
}
\usage{
cei_colour_4
}
\description{
CEI colour palette with four colours
}
\keyword{datasets}
