% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colour_6}
\alias{cei_colour_6}
\title{CEI colour palette with six colours}
\format{
list
}
\usage{
cei_colour_6
}
\description{
CEI colour palette with six colours
}
\keyword{datasets}
