% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colour_7}
\alias{cei_colour_7}
\title{CEI colour palette with seven colours}
\format{
list
}
\usage{
cei_colour_7
}
\description{
CEI colour palette with seven colours
}
\keyword{datasets}
