% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ceiglobal.R
\docType{data}
\name{cei_colour_8}
\alias{cei_colour_8}
\title{CEI colour palette with eight colours}
\format{
list
}
\usage{
cei_colour_8
}
\description{
CEI colour palette with eight colours
}
\keyword{datasets}
