% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colour_9}
\alias{cei_colour_9}
\title{CEI colour palette with nine colours}
\format{
list
}
\usage{
cei_colour_9
}
\description{
CEI colour palette with nine colours
}
\keyword{datasets}
