% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ceiglobal.R
\docType{data}
\name{cei_colour_palette}
\alias{cei_colour_palette}
\title{List of named hex codes for CEI colour palette from style guide}
\format{
list
}
\usage{
cei_colour_palette
}
\description{
List of named hex codes for CEI colour palette from style guide
}
\keyword{datasets}
