% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_colours}
\alias{cei_colours}
\title{List of hex codes for CEI colour palette from style guide}
\format{
list
}
\usage{
cei_colours
}
\arguments{
\item{cei_colours}{List of hex codes for CEI colour palette from style guide}

\item{cei_grey}{CEI grey colour from style guide}

\item{cei_colour_palette}{List of named hex codes for CEI colour palette from style guide}

\item{cei_colour_2}{two toned palette for figures using CEI style guide}

\item{cei_colour_3}{three toned palette for figures using CEI style guide}

\item{cei_colour_4}{four toned palette for figures using CEI style guide}

\item{cei_colour_5}{five toned palette for figures using CEI style guide}

\item{cei_colour_6}{six toned palette for figures using CEI style guide}

\item{cei_colour_7}{seven toned palette for figures using CEI style guide}

\item{cei_colour_8}{eight toned palette for figures using CEI style guide}

\item{cei_colour_9}{nine toned palette for figures using CEI style guide}

\item{cei_colour_10}{ten toned palette for figures using CEI style guide}

\item{cei_colour_11}{eleven toned palette for figures using CEI style guide}

\item{cei_colour_12}{twelve toned palette for figures using CEI style guide}
}
\description{
List of hex codes for CEI colour palette from style guide
}
\section{This section details the colour scheme used in \code{theme_ceiglobal}}{
NA
}

\keyword{datasets}
