% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ceiglobal.R
\docType{data}
\name{cei_dark_grey}
\alias{cei_dark_grey}
\title{CEI dark grey colour from style guide}
\format{
length 1 character vector
}
\usage{
cei_dark_grey
}
\description{
CEI dark grey colour from style guide
}
\keyword{datasets}
