% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_dark_purple}
\alias{cei_dark_purple}
\title{CEI dark purple colour from style guide}
\format{
length 1 character vector
}
\usage{
cei_dark_purple
}
\description{
CEI dark purple colour from style guide
}
\keyword{datasets}
