% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_green}
\alias{cei_green}
\title{CEI purple colour from style guide}
\format{
length 1 character vector
}
\usage{
cei_green
}
\description{
CEI purple colour from style guide
}
\keyword{datasets}
