% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_light_green}
\alias{cei_light_green}
\title{CEI light green from style guide}
\format{
length 1 character vector
}
\usage{
cei_light_green
}
\description{
CEI light green from style guide
}
\keyword{datasets}
