% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_line_grey}
\alias{cei_line_grey}
\title{CEI line grey from style guide}
\format{
length 1 character vector
}
\usage{
cei_line_grey
}
\description{
CEI line grey from style guide
}
\keyword{datasets}
