% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{data}
\name{cei_yellow}
\alias{cei_yellow}
\title{CEI yellow from style guide}
\format{
length 1 character vector
}
\usage{
cei_yellow
}
\description{
CEI yellow from style guide
}
\keyword{datasets}
