% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ceiglobal.R
\docType{data}
\name{font_raleway}
\alias{font_raleway}
\title{Raleway font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_raleway
}
\description{
\code{font_raleway} == "\code{Raleway}"
}
\keyword{datasets}
