% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\name{theme_ceiglobal}
\alias{theme_ceiglobal}
\title{Title}
\usage{
theme_ceiglobal(base_size = 10, base_family = "Raleway", ticks = TRUE)
}
\arguments{
\item{ticks}{}
}
\value{

}
\description{
Title
}
\examples{
mtcars \%>\%
ggplot() +
 aes(x = cyl, y = mpg, fill = gear) +
 geom_bar(stat = "identity") +
 facet_grid(gear ~ vs) +
 theme_ceiglobal()
}
