% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ceiglobal.R
\docType{package}
\name{ceiglobal}
\alias{ceiglobal}
\title{Additional Themes and Theme Components for 'ggplot2'}
\description{
A theme for 'ggplot2' based on the cei style guide
}
\details{
The core theme: \code{theme_ceiglobal} is awesome.
}
\author{
Dave Taylor (david.taylor@ceiglobal.org)
}
\keyword{internal}
